+++ checkpoint/engine.go
@@ -33,18 +33,18 @@ var (
 	ErrIncompatibleHashes               = errors.New("incompatible hashes")
 
 	cpOrder = []types.CheckpointName{
-		types.ValidatorsCheckpoint,      // validators information
-		types.AssetsCheckpoint,          // assets are required for collateral to work, and the vote asset needs to be restored
-		types.CollateralCheckpoint,      // without balances, governance (proposals, bonds) are difficult
-		types.NetParamsCheckpoint,       // net params should go right after assets and collateral, so vote tokens are restored
-		types.GovernanceCheckpoint,      // depends on all of the above
-		types.EpochCheckpoint,           // restore epoch information... so delegation sequence ID's make sense
-		types.MultisigControlCheckpoint, // restore the staking information, so delegation make sense
-		types.StakingCheckpoint,         // restore the staking information, so delegation make sense
+		types.ValidatorsCheckpoint,            // validators information
+		types.AssetsCheckpoint,                // assets are required for collateral to work, and the vote asset needs to be restored
+		types.CollateralCheckpoint,            // without balances, governance (proposals, bonds) are difficult
+		types.NetParamsCheckpoint,             // net params should go right after assets and collateral, so vote tokens are restored
+		types.MarketActivityTrackerCheckpoint, // restore market activity information - needs to happen before governance
+		types.GovernanceCheckpoint,            // depends on all of the above
+		types.EpochCheckpoint,                 // restore epoch information... so delegation sequence ID's make sense
+		types.MultisigControlCheckpoint,       // restore the staking information, so delegation make sense
+		types.StakingCheckpoint,               // restore the staking information, so delegation make sense
 		types.DelegationCheckpoint,
-		types.PendingRewardsCheckpoint,        // pending rewards can basically be reloaded any time
-		types.MarketActivityTrackerCheckpoint, // restore market activity information
-		types.BankingCheckpoint,               // Banking checkpoint needs to be reload any time after collateral
+		types.PendingRewardsCheckpoint, // pending rewards can basically be reloaded any time
+		types.BankingCheckpoint,        // Banking checkpoint needs to be reload any time after collateral
 
 	}
 )