+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The init command now also generate the configuration for tendermint, the flags `
 
 - [5579](https://github.com/vegaprotocol/vega/issues/5579) - Make vega a built-in Tendermint application
 - [5249](https://github.com/vegaprotocol/vega/issues/5249) - Migrate to Tendermint version 0.35.8
+- [5624](https://github.com/vegaprotocol/vega/issues/5624) - Get rid of `updateFrequency` in price monitoring definition
 
 ### 🗑️ Deprecation
 - [](https://github.com/vegaprotocol/vega/issues/xxxx) -
@@ -25,7 +26,10 @@ The init command now also generate the configuration for tendermint, the flags `
 ### 🐛 Fixes
 - [5571](https://github.com/vegaprotocol/vega/issues/5571) - Restore pending assets status correctly after snapshot restore
 - [5572](https://github.com/vegaprotocol/vega/issues/5572) - Add validation on `IDs` and public keys
+- [5348](https://github.com/vegaprotocol/vega/issues/5348) - Restore markets from checkpoint proposal
+- [5279](https://github.com/vegaprotocol/vega/issues/5279) - Fix loading of proposals from checkpoint 
 - [5598](https://github.com/vegaprotocol/vega/issues/5598) - Remove `currentTime` from topology engine to ease snapshot restoration
+- [5609](https://github.com/vegaprotocol/vega/issues/5609) - Set event forwarder last seen height after snapshot restore
 
 ## 0.53.0
 
@@ -138,6 +142,7 @@ The init command now also generate the configuration for tendermint, the flags `
 - [5268](https://github.com/vegaprotocol/vega/issues/5268) - Make validator heartbeat frequency a function of the epoch duration.
 - [5271](https://github.com/vegaprotocol/vega/issues/5271) - Make generated hex IDs lower case
 - [5273](https://github.com/vegaprotocol/vega/issues/5273) - Reward / Transfer to allow payout of reward in an arbitrary asset unrelated to the settlement and by market.
+- [5207](https://github.com/vegaprotocol/vega/issues/5206) - Add integration tests to ensure price bounds and decimal places work as expected
 - [5243](https://github.com/vegaprotocol/vega/issues/5243) - Update equity like share according to spec changes.
 - [5249](https://github.com/vegaprotocol/vega/issues/5249) - Upgrade to tendermint 0.35.6